+++ src/api/ai.js
@@ -1,16 +1,17 @@
 import { jsonFetch } from "./auth.js";
 
-export function generateCourseOutlineFromTitle(title) {
-  console.log(title);
-  return jsonFetch(`${import.meta.env.VITE_API_BASE_URL}/ai/outline/`, {
+const API = import.meta.env.VITE_API_BASE_URL;
+
+export function generateCourseOutline({ title }) {
+  return jsonFetch(`${API}/ai/outline`, {
     method: "POST",
     body: JSON.stringify({ title }),
   });
 }
 
-export function generateCoursewareFromTitle(courseTitle, courseId, title) {
-  return jsonFetch(`${import.meta.env.VITE_API_BASE_URL}/ai/courseware/`, {
+export function generateCourseware({ courseId, courseTitle, title }) {
+  return jsonFetch(`${API}/ai/courseware`, {
     method: "POST",
-    body: JSON.stringify({ courseTitle, courseId, title }),
+    body: JSON.stringify({ courseId, courseTitle, title }),
   });
 }
