--- src/mock/courses.js
@@ -1,287 +1,0 @@
-export default [
-  {
-    _id: {
-      $oid: "b9374ca8151541e99d121471",
-    },
-    title: "Course 1 Title",
-    coursewares: [
-      {
-        title: "Courseware 1 of Course 1",
-        coursewareId: {
-          $oid: "e6845359a4b244e885778d69",
-        },
-      },
-      {
-        title: "Courseware 2 of Course 1",
-        coursewareId: {
-          $oid: "32f7589f950f477298ccaf3b",
-        },
-      },
-      {
-        title: "Courseware 3 of Course 1",
-        coursewareId: {
-          $oid: "8b948422c28443b68743e7c3",
-        },
-      },
-      {
-        title: "Courseware 4 of Course 1",
-        coursewareId: {
-          $oid: "4003f178ff8b401787e09de4",
-        },
-      },
-      {
-        title: "Courseware 5 of Course 1",
-        coursewareId: {
-          $oid: "1a9a627de0ee4c63acc4aa3e",
-        },
-      },
-      {
-        title: "Courseware 6 of Course 1",
-        coursewareId: {
-          $oid: "7af4f5e452624ee0ad60f8da",
-        },
-      },
-      {
-        title: "Courseware 7 of Course 1",
-        coursewareId: {
-          $oid: "0ac113dabf1d44f98caa329a",
-        },
-      },
-    ],
-    __v: 0,
-    keywords: [],
-    stats: {
-      likes: 0,
-      dislikes: 0,
-      views: 0,
-    },
-  },
-  {
-    _id: {
-      $oid: "709ac14d9cb84698a0720725",
-    },
-    title: "Course 2 Title",
-    coursewares: [
-      {
-        title: "Courseware 1 of Course 2",
-        coursewareId: {
-          $oid: "ac033b12bf454ee39b21bcd4",
-        },
-      },
-      {
-        title: "Courseware 2 of Course 2",
-        coursewareId: {
-          $oid: "641d9a9d1515484dbc594694",
-        },
-      },
-      {
-        title: "Courseware 3 of Course 2",
-        coursewareId: {
-          $oid: "b6f9ddb28021471ab0be0b9e",
-        },
-      },
-      {
-        title: "Courseware 4 of Course 2",
-        coursewareId: {
-          $oid: "6d9c23e2864640a2a7e76392",
-        },
-      },
-      {
-        title: "Courseware 5 of Course 2",
-        coursewareId: {
-          $oid: "c0d4d583bab541b9b55a5f45",
-        },
-      },
-      {
-        title: "Courseware 6 of Course 2",
-        coursewareId: {
-          $oid: "51d6f8068e634d528c7ea322",
-        },
-      },
-      {
-        title: "Courseware 7 of Course 2",
-        coursewareId: {
-          $oid: "36696224d96a42668217f050",
-        },
-      },
-    ],
-    __v: 0,
-    keywords: [],
-    stats: {
-      likes: 0,
-      dislikes: 0,
-      views: 0,
-    },
-  },
-  {
-    _id: {
-      $oid: "a1e1084d87dc4502a1dba5f5",
-    },
-    title: "Course 3 Title",
-    coursewares: [
-      {
-        title: "Courseware 1 of Course 3",
-        coursewareId: {
-          $oid: "b1f1058b5caf4ffdb44cb516",
-        },
-      },
-      {
-        title: "Courseware 2 of Course 3",
-        coursewareId: {
-          $oid: "fda7201018324f5e8dbc9379",
-        },
-      },
-      {
-        title: "Courseware 3 of Course 3",
-        coursewareId: {
-          $oid: "7e24dd8e414a43a5b393c11d",
-        },
-      },
-      {
-        title: "Courseware 4 of Course 3",
-        coursewareId: {
-          $oid: "adc2b4e649c546549f13c006",
-        },
-      },
-      {
-        title: "Courseware 5 of Course 3",
-        coursewareId: {
-          $oid: "af11c9563b584818ae9f2090",
-        },
-      },
-      {
-        title: "Courseware 6 of Course 3",
-        coursewareId: {
-          $oid: "42dc3829c98741348847554e",
-        },
-      },
-      {
-        title: "Courseware 7 of Course 3",
-        coursewareId: {
-          $oid: "bfb4701e1eb34c51a6ab747b",
-        },
-      },
-    ],
-    __v: 0,
-    keywords: [],
-    stats: {
-      likes: 0,
-      dislikes: 0,
-      views: 0,
-    },
-  },
-  {
-    _id: {
-      $oid: "b4b2a2a432ad4c5f82b76136",
-    },
-    title: "Course 4 Title",
-    coursewares: [
-      {
-        title: "Courseware 1 of Course 4",
-        coursewareId: {
-          $oid: "7817bb1d1f754899bb0de30c",
-        },
-      },
-      {
-        title: "Courseware 2 of Course 4",
-        coursewareId: {
-          $oid: "6be1f5ca22d74c9abf334b66",
-        },
-      },
-      {
-        title: "Courseware 3 of Course 4",
-        coursewareId: {
-          $oid: "203394bf8af144b2800d65aa",
-        },
-      },
-      {
-        title: "Courseware 4 of Course 4",
-        coursewareId: {
-          $oid: "eb1321bd92694cbea66b3b4d",
-        },
-      },
-      {
-        title: "Courseware 5 of Course 4",
-        coursewareId: {
-          $oid: "29dfc7ee88f44e1681ed5617",
-        },
-      },
-      {
-        title: "Courseware 6 of Course 4",
-        coursewareId: {
-          $oid: "d69b5965c102446ba2acce14",
-        },
-      },
-      {
-        title: "Courseware 7 of Course 4",
-        coursewareId: {
-          $oid: "eeac0566a0a04785b25af656",
-        },
-      },
-    ],
-    __v: 0,
-    keywords: [],
-    stats: {
-      likes: 0,
-      dislikes: 0,
-      views: 0,
-    },
-  },
-  {
-    _id: {
-      $oid: "dfc694ce06de4034b1f71534",
-    },
-    title: "Course 5 Title",
-    coursewares: [
-      {
-        title: "Courseware 1 of Course 5",
-        coursewareId: {
-          $oid: "7ba27343c11643a2b9e1c53a",
-        },
-      },
-      {
-        title: "Courseware 2 of Course 5",
-        coursewareId: {
-          $oid: "46c4520c6b67439d9b8f016a",
-        },
-      },
-      {
-        title: "Courseware 3 of Course 5",
-        coursewareId: {
-          $oid: "f7777261ea95418c91c69f78",
-        },
-      },
-      {
-        title: "Courseware 4 of Course 5",
-        coursewareId: {
-          $oid: "c1ac496500164318b929c985",
-        },
-      },
-      {
-        title: "Courseware 5 of Course 5",
-        coursewareId: {
-          $oid: "e836fd3cdce24c248a458f66",
-        },
-      },
-      {
-        title: "Courseware 6 of Course 5",
-        coursewareId: {
-          $oid: "f9438773d98d408cb973f031",
-        },
-      },
-      {
-        title: "Courseware 7 of Course 5",
-        coursewareId: {
-          $oid: "ab419547702a4783a90b8520",
-        },
-      },
-    ],
-    __v: 0,
-    keywords: [],
-    stats: {
-      likes: 0,
-      dislikes: 0,
-      views: 0,
-    },
-  },
-];
