--- src/mock/questions.js
@@ -1,2627 +1,0 @@
-export const mockQuestions = [
-  {
-    "_id": {
-      "$oid": "56f44081a6b049aca39a4ef9"
-    },
-    "coursewareId": {
-      "$oid": "e6845359a4b244e885778d69"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 1 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 1 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "8a90a9a22e0142939e1b15d5"
-    },
-    "coursewareId": {
-      "$oid": "e6845359a4b244e885778d69"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 1 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 1 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "98394d3e2e1f4bbb926d2c85"
-    },
-    "coursewareId": {
-      "$oid": "e6845359a4b244e885778d69"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 1 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 1 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "7bbb6987b9944156bf99bc8a"
-    },
-    "coursewareId": {
-      "$oid": "e6845359a4b244e885778d69"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 1 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 1 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "4601736fb68c47c1a0eee741"
-    },
-    "coursewareId": {
-      "$oid": "e6845359a4b244e885778d69"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 1 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 1 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "6e0a4955d3c14053a67be568"
-    },
-    "coursewareId": {
-      "$oid": "32f7589f950f477298ccaf3b"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 2 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 2 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "c3b4461a58e14dd1a0efd117"
-    },
-    "coursewareId": {
-      "$oid": "32f7589f950f477298ccaf3b"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 2 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 2 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "2a4d47f449f04a8e98fc46ba"
-    },
-    "coursewareId": {
-      "$oid": "32f7589f950f477298ccaf3b"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 2 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 2 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "1a30c7aac77d40f4858c121c"
-    },
-    "coursewareId": {
-      "$oid": "32f7589f950f477298ccaf3b"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 2 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 2 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "2478214d7ec64fa380b5963c"
-    },
-    "coursewareId": {
-      "$oid": "32f7589f950f477298ccaf3b"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 2 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 2 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "ecc40bcf1bde4f3790dc9d02"
-    },
-    "coursewareId": {
-      "$oid": "8b948422c28443b68743e7c3"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 3 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 3 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "58f3619159774523bf648fbb"
-    },
-    "coursewareId": {
-      "$oid": "8b948422c28443b68743e7c3"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 3 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 3 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "be4150c3e30840798f60a121"
-    },
-    "coursewareId": {
-      "$oid": "8b948422c28443b68743e7c3"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 3 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 3 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "bb57fc18bf974b4893b3332a"
-    },
-    "coursewareId": {
-      "$oid": "8b948422c28443b68743e7c3"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 3 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 3 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "33b7219c54a6487086c4df42"
-    },
-    "coursewareId": {
-      "$oid": "8b948422c28443b68743e7c3"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 3 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 3 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "fde3d813803f489e98a23e18"
-    },
-    "coursewareId": {
-      "$oid": "4003f178ff8b401787e09de4"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 4 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 4 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "e1508656465440279c98cfa6"
-    },
-    "coursewareId": {
-      "$oid": "4003f178ff8b401787e09de4"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 4 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 4 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "4ebd0c2f0bf843e8a82795a1"
-    },
-    "coursewareId": {
-      "$oid": "4003f178ff8b401787e09de4"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 4 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 4 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "248bf5033ae14eab9df00907"
-    },
-    "coursewareId": {
-      "$oid": "4003f178ff8b401787e09de4"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 4 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 4 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "9ba42094132d48f993e5e6fb"
-    },
-    "coursewareId": {
-      "$oid": "4003f178ff8b401787e09de4"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 4 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 4 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "a4717215d1f14fd49f6c9049"
-    },
-    "coursewareId": {
-      "$oid": "1a9a627de0ee4c63acc4aa3e"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 5 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 5 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "4c9f59c1ac13411d99b361e1"
-    },
-    "coursewareId": {
-      "$oid": "1a9a627de0ee4c63acc4aa3e"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 5 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 5 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "c18d5d84031c470c9fced745"
-    },
-    "coursewareId": {
-      "$oid": "1a9a627de0ee4c63acc4aa3e"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 5 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 5 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "e2a86166aa70436e8ee9752e"
-    },
-    "coursewareId": {
-      "$oid": "1a9a627de0ee4c63acc4aa3e"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 5 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 5 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "4b885e93d6a34bb0a26e7024"
-    },
-    "coursewareId": {
-      "$oid": "1a9a627de0ee4c63acc4aa3e"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 5 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 5 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "ca43620632b0445980e522b4"
-    },
-    "coursewareId": {
-      "$oid": "7af4f5e452624ee0ad60f8da"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 6 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 6 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "67903ca3d3e24b5ba1dd1b98"
-    },
-    "coursewareId": {
-      "$oid": "7af4f5e452624ee0ad60f8da"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 6 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 6 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "042f7f26ed9f4632a4867c37"
-    },
-    "coursewareId": {
-      "$oid": "7af4f5e452624ee0ad60f8da"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 6 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 6 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "c95e881adccd45f69b774034"
-    },
-    "coursewareId": {
-      "$oid": "7af4f5e452624ee0ad60f8da"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 6 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 6 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "b9e6adea9acb4743b34376ab"
-    },
-    "coursewareId": {
-      "$oid": "7af4f5e452624ee0ad60f8da"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 6 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 6 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "22c40a2994ef418789632090"
-    },
-    "coursewareId": {
-      "$oid": "0ac113dabf1d44f98caa329a"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 7 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 7 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "165fb57f5ab54eaf82341c79"
-    },
-    "coursewareId": {
-      "$oid": "0ac113dabf1d44f98caa329a"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 7 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 7 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "cab8f4b3472d49c78640b6b2"
-    },
-    "coursewareId": {
-      "$oid": "0ac113dabf1d44f98caa329a"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 7 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 7 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "343387fddbea49658c9a6f5b"
-    },
-    "coursewareId": {
-      "$oid": "0ac113dabf1d44f98caa329a"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 7 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 7 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "15b3a832458c473a9fe77430"
-    },
-    "coursewareId": {
-      "$oid": "0ac113dabf1d44f98caa329a"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 7 of Course 1",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 7 of Course 1?"
-  },
-  {
-    "_id": {
-      "$oid": "6af323b45c20433086e5b88f"
-    },
-    "coursewareId": {
-      "$oid": "ac033b12bf454ee39b21bcd4"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 1 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 1 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "9b66d6dd34fc4ac687dd1239"
-    },
-    "coursewareId": {
-      "$oid": "ac033b12bf454ee39b21bcd4"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 1 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 1 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "abe551294f6747b9a732d1cc"
-    },
-    "coursewareId": {
-      "$oid": "ac033b12bf454ee39b21bcd4"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 1 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 1 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "40caca9f01b64c718b7e9ffb"
-    },
-    "coursewareId": {
-      "$oid": "ac033b12bf454ee39b21bcd4"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 1 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 1 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "c67f0f74c1024f5eb48635c9"
-    },
-    "coursewareId": {
-      "$oid": "ac033b12bf454ee39b21bcd4"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 1 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 1 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "cd8ed32ae20e4c63a5a516a5"
-    },
-    "coursewareId": {
-      "$oid": "641d9a9d1515484dbc594694"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 2 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 2 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "8590b534c469432997e983fb"
-    },
-    "coursewareId": {
-      "$oid": "641d9a9d1515484dbc594694"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 2 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 2 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "b247488de11d4b2d88b58d41"
-    },
-    "coursewareId": {
-      "$oid": "641d9a9d1515484dbc594694"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 2 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 2 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "aae06241e5fe43ac99ca41bd"
-    },
-    "coursewareId": {
-      "$oid": "641d9a9d1515484dbc594694"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 2 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 2 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "2579f40ec6fa4232a3f88a26"
-    },
-    "coursewareId": {
-      "$oid": "641d9a9d1515484dbc594694"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 2 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 2 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "e961217abe954f7b8b395282"
-    },
-    "coursewareId": {
-      "$oid": "b6f9ddb28021471ab0be0b9e"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 3 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 3 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "ddddad8ceb3a4fe9a4ba8ef9"
-    },
-    "coursewareId": {
-      "$oid": "b6f9ddb28021471ab0be0b9e"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 3 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 3 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "b4cafcf89a724fb49161e158"
-    },
-    "coursewareId": {
-      "$oid": "b6f9ddb28021471ab0be0b9e"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 3 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 3 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "1029896111c541ddb85bb661"
-    },
-    "coursewareId": {
-      "$oid": "b6f9ddb28021471ab0be0b9e"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 3 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 3 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "2de31f00c3214b7c9f08f032"
-    },
-    "coursewareId": {
-      "$oid": "b6f9ddb28021471ab0be0b9e"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 3 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 3 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "fb273c9cca4d481ab9ef7fa9"
-    },
-    "coursewareId": {
-      "$oid": "6d9c23e2864640a2a7e76392"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 4 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 4 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "290fb975f4804d63b60227e9"
-    },
-    "coursewareId": {
-      "$oid": "6d9c23e2864640a2a7e76392"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 4 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 4 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "9ed6ab2dde9e49519d712a93"
-    },
-    "coursewareId": {
-      "$oid": "6d9c23e2864640a2a7e76392"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 4 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 4 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "15c2e112117d4cf9961b86a3"
-    },
-    "coursewareId": {
-      "$oid": "6d9c23e2864640a2a7e76392"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 4 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 4 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "43bec193a79c4b6b90ec5d17"
-    },
-    "coursewareId": {
-      "$oid": "6d9c23e2864640a2a7e76392"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 4 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 4 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "22015c8ebcd24fa8a263b1c8"
-    },
-    "coursewareId": {
-      "$oid": "c0d4d583bab541b9b55a5f45"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 5 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 5 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "334b135bba9d4a5dbeff93e1"
-    },
-    "coursewareId": {
-      "$oid": "c0d4d583bab541b9b55a5f45"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 5 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 5 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "56cf7d59ab0b4f1193816b5a"
-    },
-    "coursewareId": {
-      "$oid": "c0d4d583bab541b9b55a5f45"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 5 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 5 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "69d88271f1aa4b979a819fe0"
-    },
-    "coursewareId": {
-      "$oid": "c0d4d583bab541b9b55a5f45"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 5 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 5 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "65fcd28a460e49d5a03e5cdd"
-    },
-    "coursewareId": {
-      "$oid": "c0d4d583bab541b9b55a5f45"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 5 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 5 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "29d502b20b6c4bd3a1920eef"
-    },
-    "coursewareId": {
-      "$oid": "51d6f8068e634d528c7ea322"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 6 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 6 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "9d4dce4865af4455b1f79c56"
-    },
-    "coursewareId": {
-      "$oid": "51d6f8068e634d528c7ea322"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 6 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 6 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "ad2b69317d474c87860cd6f0"
-    },
-    "coursewareId": {
-      "$oid": "51d6f8068e634d528c7ea322"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 6 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 6 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "91d55cefe56f4192b310e6c9"
-    },
-    "coursewareId": {
-      "$oid": "51d6f8068e634d528c7ea322"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 6 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 6 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "d89c2b19a3a04801ad4adfde"
-    },
-    "coursewareId": {
-      "$oid": "51d6f8068e634d528c7ea322"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 6 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 6 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "d7682d07162841e68c2561ed"
-    },
-    "coursewareId": {
-      "$oid": "36696224d96a42668217f050"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 7 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 7 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "d56d332c7480470889e727aa"
-    },
-    "coursewareId": {
-      "$oid": "36696224d96a42668217f050"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 7 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 7 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "594c73490e7c4c458fc243f6"
-    },
-    "coursewareId": {
-      "$oid": "36696224d96a42668217f050"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 7 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 7 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "09fd590e147842b3a9bf9757"
-    },
-    "coursewareId": {
-      "$oid": "36696224d96a42668217f050"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 7 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 7 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "182e2dae16a345a4af60f26b"
-    },
-    "coursewareId": {
-      "$oid": "36696224d96a42668217f050"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 7 of Course 2",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 7 of Course 2?"
-  },
-  {
-    "_id": {
-      "$oid": "20b5d8630e624ed68cbd9bc5"
-    },
-    "coursewareId": {
-      "$oid": "b1f1058b5caf4ffdb44cb516"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 1 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 1 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "8470ab7011a74cd58047eacf"
-    },
-    "coursewareId": {
-      "$oid": "b1f1058b5caf4ffdb44cb516"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 1 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 1 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "c5a91983b47346009faadc8e"
-    },
-    "coursewareId": {
-      "$oid": "b1f1058b5caf4ffdb44cb516"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 1 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 1 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "93db6467da4545abb3d6d225"
-    },
-    "coursewareId": {
-      "$oid": "b1f1058b5caf4ffdb44cb516"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 1 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 1 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "95f25b05eb1e4108848e9af5"
-    },
-    "coursewareId": {
-      "$oid": "b1f1058b5caf4ffdb44cb516"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 1 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 1 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "1205d2f300b941988aa1b1d5"
-    },
-    "coursewareId": {
-      "$oid": "fda7201018324f5e8dbc9379"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 2 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 2 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "357384991a60454fa16d0cef"
-    },
-    "coursewareId": {
-      "$oid": "fda7201018324f5e8dbc9379"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 2 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 2 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "7490b3ca881f4f86acdb04d0"
-    },
-    "coursewareId": {
-      "$oid": "fda7201018324f5e8dbc9379"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 2 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 2 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "b02d4aba8caf4e9bb8a939d6"
-    },
-    "coursewareId": {
-      "$oid": "fda7201018324f5e8dbc9379"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 2 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 2 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "f294c6cd08364c31b6f511d6"
-    },
-    "coursewareId": {
-      "$oid": "fda7201018324f5e8dbc9379"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 2 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 2 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "ded8721e9ad743ce9a5f8518"
-    },
-    "coursewareId": {
-      "$oid": "7e24dd8e414a43a5b393c11d"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 3 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 3 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "15a479c5154c489cb82ea435"
-    },
-    "coursewareId": {
-      "$oid": "7e24dd8e414a43a5b393c11d"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 3 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 3 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "3ffb26185967423d9349500e"
-    },
-    "coursewareId": {
-      "$oid": "7e24dd8e414a43a5b393c11d"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 3 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 3 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "04530ef1540c4412a81633a5"
-    },
-    "coursewareId": {
-      "$oid": "7e24dd8e414a43a5b393c11d"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 3 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 3 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "272d5fa6661a4be486d0f4a6"
-    },
-    "coursewareId": {
-      "$oid": "7e24dd8e414a43a5b393c11d"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 3 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 3 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "5a028963125d43e88748df14"
-    },
-    "coursewareId": {
-      "$oid": "adc2b4e649c546549f13c006"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 4 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 4 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "5c83b6f2e8d84a4683a939c1"
-    },
-    "coursewareId": {
-      "$oid": "adc2b4e649c546549f13c006"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 4 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 4 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "b570fe7f610943309469b6cf"
-    },
-    "coursewareId": {
-      "$oid": "adc2b4e649c546549f13c006"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 4 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 4 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "464670513ad3419b9f1250d2"
-    },
-    "coursewareId": {
-      "$oid": "adc2b4e649c546549f13c006"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 4 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 4 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "ad8be3adcf274f6eaba9020d"
-    },
-    "coursewareId": {
-      "$oid": "adc2b4e649c546549f13c006"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 4 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 4 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "e1ff590fa6854240aad03cdc"
-    },
-    "coursewareId": {
-      "$oid": "af11c9563b584818ae9f2090"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 5 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 5 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "0b85fb7ec71044a1a0ece1b8"
-    },
-    "coursewareId": {
-      "$oid": "af11c9563b584818ae9f2090"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 5 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 5 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "355fd6b83b7e42bea5baefc6"
-    },
-    "coursewareId": {
-      "$oid": "af11c9563b584818ae9f2090"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 5 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 5 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "c19e55b3b7c945cc9daa4edd"
-    },
-    "coursewareId": {
-      "$oid": "af11c9563b584818ae9f2090"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 5 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 5 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "83db795527944ed0973792ed"
-    },
-    "coursewareId": {
-      "$oid": "af11c9563b584818ae9f2090"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 5 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 5 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "c7e52872445348aaafc4eac7"
-    },
-    "coursewareId": {
-      "$oid": "42dc3829c98741348847554e"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 6 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 6 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "f7298618d4a84bb2a95998f6"
-    },
-    "coursewareId": {
-      "$oid": "42dc3829c98741348847554e"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 6 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 6 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "e3ae1d08b5874660971fc53a"
-    },
-    "coursewareId": {
-      "$oid": "42dc3829c98741348847554e"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 6 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 6 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "e0cbe0c4dddd42668d006121"
-    },
-    "coursewareId": {
-      "$oid": "42dc3829c98741348847554e"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 6 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 6 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "79b2a133feab4fd1a03e70f6"
-    },
-    "coursewareId": {
-      "$oid": "42dc3829c98741348847554e"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 6 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 6 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "a355a389706c491da3f97bd0"
-    },
-    "coursewareId": {
-      "$oid": "bfb4701e1eb34c51a6ab747b"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 7 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 7 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "8367f813ed7e40ac8849af84"
-    },
-    "coursewareId": {
-      "$oid": "bfb4701e1eb34c51a6ab747b"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 7 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 7 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "ceb971f414244879be1639e4"
-    },
-    "coursewareId": {
-      "$oid": "bfb4701e1eb34c51a6ab747b"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 7 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 7 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "1197909e31dd41b781372c4f"
-    },
-    "coursewareId": {
-      "$oid": "bfb4701e1eb34c51a6ab747b"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 7 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 7 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "f0fa0ae314684707a6a7fda1"
-    },
-    "coursewareId": {
-      "$oid": "bfb4701e1eb34c51a6ab747b"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 7 of Course 3",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 7 of Course 3?"
-  },
-  {
-    "_id": {
-      "$oid": "dc16e85d238e40f2883f4210"
-    },
-    "coursewareId": {
-      "$oid": "7817bb1d1f754899bb0de30c"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 1 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 1 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "3e0fa5abbf124ebd8f53636f"
-    },
-    "coursewareId": {
-      "$oid": "7817bb1d1f754899bb0de30c"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 1 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 1 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "154d79e4607341269b6985e0"
-    },
-    "coursewareId": {
-      "$oid": "7817bb1d1f754899bb0de30c"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 1 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 1 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "5bf85fe7f4994ae4ae3d537c"
-    },
-    "coursewareId": {
-      "$oid": "7817bb1d1f754899bb0de30c"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 1 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 1 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "923da2aef0b84bf9a13ea26f"
-    },
-    "coursewareId": {
-      "$oid": "7817bb1d1f754899bb0de30c"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 1 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 1 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "3eba936789074eaf8c90509f"
-    },
-    "coursewareId": {
-      "$oid": "6be1f5ca22d74c9abf334b66"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 2 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 2 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "f5905249c919430e8007ebcb"
-    },
-    "coursewareId": {
-      "$oid": "6be1f5ca22d74c9abf334b66"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 2 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 2 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "3441ac4d0e5b402ba55c12e6"
-    },
-    "coursewareId": {
-      "$oid": "6be1f5ca22d74c9abf334b66"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 2 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 2 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "4b83f76e8a744eba9a1c4157"
-    },
-    "coursewareId": {
-      "$oid": "6be1f5ca22d74c9abf334b66"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 2 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 2 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "15767be3a1ec4cf28b27efc8"
-    },
-    "coursewareId": {
-      "$oid": "6be1f5ca22d74c9abf334b66"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 2 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 2 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "80d0be262f234afaadc270ae"
-    },
-    "coursewareId": {
-      "$oid": "203394bf8af144b2800d65aa"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 3 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 3 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "8f5fe8a543a94f9897ad4ee2"
-    },
-    "coursewareId": {
-      "$oid": "203394bf8af144b2800d65aa"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 3 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 3 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "b6601b19369f40ff95d8872a"
-    },
-    "coursewareId": {
-      "$oid": "203394bf8af144b2800d65aa"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 3 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 3 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "bd10f1fcbbe9441eb4600ffe"
-    },
-    "coursewareId": {
-      "$oid": "203394bf8af144b2800d65aa"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 3 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 3 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "499d8dddfe0e48fca8c7ed65"
-    },
-    "coursewareId": {
-      "$oid": "203394bf8af144b2800d65aa"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 3 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 3 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "8ef7160f2b9b4ea49f441ab0"
-    },
-    "coursewareId": {
-      "$oid": "eb1321bd92694cbea66b3b4d"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 4 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 4 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "1995d3bd73e9448f94fc2b2d"
-    },
-    "coursewareId": {
-      "$oid": "eb1321bd92694cbea66b3b4d"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 4 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 4 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "377ff5a32ee34e5c8ca3680b"
-    },
-    "coursewareId": {
-      "$oid": "eb1321bd92694cbea66b3b4d"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 4 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 4 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "de52bd427bcb45deb595d71d"
-    },
-    "coursewareId": {
-      "$oid": "eb1321bd92694cbea66b3b4d"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 4 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 4 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "f0439da290eb4dceb2918710"
-    },
-    "coursewareId": {
-      "$oid": "eb1321bd92694cbea66b3b4d"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 4 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 4 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "a5330fd97b1e4e8daf6e26e1"
-    },
-    "coursewareId": {
-      "$oid": "29dfc7ee88f44e1681ed5617"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 5 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 5 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "e707f4cacb344dee83de4649"
-    },
-    "coursewareId": {
-      "$oid": "29dfc7ee88f44e1681ed5617"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 5 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 5 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "2f43d495817043cdb115a846"
-    },
-    "coursewareId": {
-      "$oid": "29dfc7ee88f44e1681ed5617"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 5 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 5 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "169260c5e902470ba68e4b91"
-    },
-    "coursewareId": {
-      "$oid": "29dfc7ee88f44e1681ed5617"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 5 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 5 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "c8c85205ae2748c0b5ad5341"
-    },
-    "coursewareId": {
-      "$oid": "29dfc7ee88f44e1681ed5617"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 5 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 5 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "336824ec9cdc4e6b9c55e82c"
-    },
-    "coursewareId": {
-      "$oid": "d69b5965c102446ba2acce14"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 6 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 6 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "1a079a8fc9884e27adb9bd0d"
-    },
-    "coursewareId": {
-      "$oid": "d69b5965c102446ba2acce14"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 6 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 6 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "7036301e7d46412285c24f12"
-    },
-    "coursewareId": {
-      "$oid": "d69b5965c102446ba2acce14"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 6 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 6 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "74f9a140de7f474f807fed57"
-    },
-    "coursewareId": {
-      "$oid": "d69b5965c102446ba2acce14"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 6 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 6 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "3b5b74b0706d49d0952beee3"
-    },
-    "coursewareId": {
-      "$oid": "d69b5965c102446ba2acce14"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 6 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 6 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "7474146229374edb92389723"
-    },
-    "coursewareId": {
-      "$oid": "eeac0566a0a04785b25af656"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 7 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 7 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "1c0063faecac40b7902fcf10"
-    },
-    "coursewareId": {
-      "$oid": "eeac0566a0a04785b25af656"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 7 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 7 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "2c1d2efd6d074451b31e7d21"
-    },
-    "coursewareId": {
-      "$oid": "eeac0566a0a04785b25af656"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 7 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 7 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "c9c2f30648c8409d9bcaf35f"
-    },
-    "coursewareId": {
-      "$oid": "eeac0566a0a04785b25af656"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 7 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 7 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "7a6b1a07a422464abc2581e7"
-    },
-    "coursewareId": {
-      "$oid": "eeac0566a0a04785b25af656"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 7 of Course 4",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 7 of Course 4?"
-  },
-  {
-    "_id": {
-      "$oid": "ab2b28e766324c93b420da9a"
-    },
-    "coursewareId": {
-      "$oid": "7ba27343c11643a2b9e1c53a"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 1 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 1 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "65b1b28eb5404c5d85a93516"
-    },
-    "coursewareId": {
-      "$oid": "7ba27343c11643a2b9e1c53a"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 1 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 1 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "acb5b51ba24a4548b4c98e6c"
-    },
-    "coursewareId": {
-      "$oid": "7ba27343c11643a2b9e1c53a"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 1 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 1 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "43d2a121e0ac494e8d2620c6"
-    },
-    "coursewareId": {
-      "$oid": "7ba27343c11643a2b9e1c53a"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 1 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 1 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "83d852c6e166410f9f43e278"
-    },
-    "coursewareId": {
-      "$oid": "7ba27343c11643a2b9e1c53a"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 1 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 1 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "a50c5a3b142f4dbd91d4e546"
-    },
-    "coursewareId": {
-      "$oid": "46c4520c6b67439d9b8f016a"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 2 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 2 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "e20db8d92de44434a8af1c8d"
-    },
-    "coursewareId": {
-      "$oid": "46c4520c6b67439d9b8f016a"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 2 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 2 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "fa7b879c61ef4466b6c50109"
-    },
-    "coursewareId": {
-      "$oid": "46c4520c6b67439d9b8f016a"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 2 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 2 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "3c0aebb6dd9242aa862e4aea"
-    },
-    "coursewareId": {
-      "$oid": "46c4520c6b67439d9b8f016a"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 2 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 2 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "0ff1984009c84a398ef52d9f"
-    },
-    "coursewareId": {
-      "$oid": "46c4520c6b67439d9b8f016a"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 2 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 2 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "1fad21a649ea460d981640de"
-    },
-    "coursewareId": {
-      "$oid": "f7777261ea95418c91c69f78"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 3 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 3 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "7bb6a358130d4dd78fe814fd"
-    },
-    "coursewareId": {
-      "$oid": "f7777261ea95418c91c69f78"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 3 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 3 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "30bad2aadf614ce8bffeb11b"
-    },
-    "coursewareId": {
-      "$oid": "f7777261ea95418c91c69f78"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 3 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 3 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "2e13fd20945f49fd8097aadc"
-    },
-    "coursewareId": {
-      "$oid": "f7777261ea95418c91c69f78"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 3 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 3 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "ca7b398094724600a0e4068b"
-    },
-    "coursewareId": {
-      "$oid": "f7777261ea95418c91c69f78"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 3 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 3 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "9767cc50e4ec4eff9ee2ace0"
-    },
-    "coursewareId": {
-      "$oid": "c1ac496500164318b929c985"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 4 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 4 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "f1001a78be2a49daa407ec9e"
-    },
-    "coursewareId": {
-      "$oid": "c1ac496500164318b929c985"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 4 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 4 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "cf6553309ecf46ea9fd5ab93"
-    },
-    "coursewareId": {
-      "$oid": "c1ac496500164318b929c985"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 4 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 4 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "862fd6ba27004aed915bcd1a"
-    },
-    "coursewareId": {
-      "$oid": "c1ac496500164318b929c985"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 4 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 4 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "766fc69584494a458edc6bf1"
-    },
-    "coursewareId": {
-      "$oid": "c1ac496500164318b929c985"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 4 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 4 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "deaa9fc186a945f0a6ada70d"
-    },
-    "coursewareId": {
-      "$oid": "e836fd3cdce24c248a458f66"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 5 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 5 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "8a68ebecbb9b40058cedff41"
-    },
-    "coursewareId": {
-      "$oid": "e836fd3cdce24c248a458f66"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 5 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 5 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "28df4a01b19b44e9a41f0bca"
-    },
-    "coursewareId": {
-      "$oid": "e836fd3cdce24c248a458f66"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 5 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 5 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "260fd7b9c21545fa991bb0b6"
-    },
-    "coursewareId": {
-      "$oid": "e836fd3cdce24c248a458f66"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 5 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 5 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "2249295d0eb640ab8a2cd7f1"
-    },
-    "coursewareId": {
-      "$oid": "e836fd3cdce24c248a458f66"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 5 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 5 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "a854a15c74b842d9b7b83eb7"
-    },
-    "coursewareId": {
-      "$oid": "f9438773d98d408cb973f031"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 6 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 6 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "478eca4dd9364bc4be8958b0"
-    },
-    "coursewareId": {
-      "$oid": "f9438773d98d408cb973f031"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 6 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 6 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "360b7b4bf38144e8a4a34195"
-    },
-    "coursewareId": {
-      "$oid": "f9438773d98d408cb973f031"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 6 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 6 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "2eeca7f0966d4b98b93ee353"
-    },
-    "coursewareId": {
-      "$oid": "f9438773d98d408cb973f031"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 6 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 6 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "e19f23f84b81497d8e87fb0a"
-    },
-    "coursewareId": {
-      "$oid": "f9438773d98d408cb973f031"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 6 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 6 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "3a5005952f6f4c6b8c2ddb5c"
-    },
-    "coursewareId": {
-      "$oid": "ab419547702a4783a90b8520"
-    },
-    "correctAnswer": "Correct Answer 1 for Courseware 7 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 1",
-      "Incorrect B 1",
-      "Incorrect C 1"
-    ],
-    "questionText": "Question 1 of Courseware 7 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "064af10022ff4830b9e21bb8"
-    },
-    "coursewareId": {
-      "$oid": "ab419547702a4783a90b8520"
-    },
-    "correctAnswer": "Correct Answer 2 for Courseware 7 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 2",
-      "Incorrect B 2",
-      "Incorrect C 2"
-    ],
-    "questionText": "Question 2 of Courseware 7 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "8a35ee1a60a84f7a94754cc7"
-    },
-    "coursewareId": {
-      "$oid": "ab419547702a4783a90b8520"
-    },
-    "correctAnswer": "Correct Answer 3 for Courseware 7 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 3",
-      "Incorrect B 3",
-      "Incorrect C 3"
-    ],
-    "questionText": "Question 3 of Courseware 7 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "8063f544a1364df8907e4c0b"
-    },
-    "coursewareId": {
-      "$oid": "ab419547702a4783a90b8520"
-    },
-    "correctAnswer": "Correct Answer 4 for Courseware 7 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 4",
-      "Incorrect B 4",
-      "Incorrect C 4"
-    ],
-    "questionText": "Question 4 of Courseware 7 of Course 5?"
-  },
-  {
-    "_id": {
-      "$oid": "2079cbcc27984a868a96a25a"
-    },
-    "coursewareId": {
-      "$oid": "ab419547702a4783a90b8520"
-    },
-    "correctAnswer": "Correct Answer 5 for Courseware 7 of Course 5",
-    "incorrectAnswers": [
-      "Incorrect A 5",
-      "Incorrect B 5",
-      "Incorrect C 5"
-    ],
-    "questionText": "Question 5 of Courseware 7 of Course 5?"
-  }
-]
