--- src/mock/users.js
@@ -1,1012 +1,0 @@
-import { mockReviewCards } from "./reviewCards.js";
-import { mockQuestions } from "./questions.js";
-
-export const mockUsers = [
-  {
-    _id: {
-      $oid: "b01342dc577445bc9800bbb6",
-    },
-    name: "User 1",
-    email: "user1@example.com",
-    password: "$2b$10$examplehash",
-    myReviewCards: [
-      {
-        _id: {
-          $oid: "13e01626d8e549ba93b64b98",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "641d9a9d1515484dbc594694",
-        },
-        questionId: {
-          $oid: "8590b534c469432997e983fb",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 2,
-        successes: 0,
-      },
-      {
-        _id: {
-          $oid: "6cb76ec9921745a6a79c3199",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "c0d4d583bab541b9b55a5f45",
-        },
-        questionId: {
-          $oid: "65fcd28a460e49d5a03e5cdd",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 7,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "80b35e767c1a4edb8e074bf4",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "ac033b12bf454ee39b21bcd4",
-        },
-        questionId: {
-          $oid: "40caca9f01b64c718b7e9ffb",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 4,
-        successes: 0,
-      },
-      {
-        _id: {
-          $oid: "6f49f60b52e2482b8dfebcb1",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "c0d4d583bab541b9b55a5f45",
-        },
-        questionId: {
-          $oid: "56cf7d59ab0b4f1193816b5a",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 9,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "2db97f58b3714f8db6059406",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "7af4f5e452624ee0ad60f8da",
-        },
-        questionId: {
-          $oid: "b9e6adea9acb4743b34376ab",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 7,
-        successes: 1,
-      },
-      {
-        _id: {
-          $oid: "28d043dae24c455f926652e9",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "641d9a9d1515484dbc594694",
-        },
-        questionId: {
-          $oid: "aae06241e5fe43ac99ca41bd",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 9,
-        successes: 0,
-      },
-      {
-        _id: {
-          $oid: "bf63e661fb5544f2a5c021e9",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "4003f178ff8b401787e09de4",
-        },
-        questionId: {
-          $oid: "9ba42094132d48f993e5e6fb",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 8,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "e1329457b1d2495b8e110c7e",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "8b948422c28443b68743e7c3",
-        },
-        questionId: {
-          $oid: "ecc40bcf1bde4f3790dc9d02",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 1,
-      },
-      {
-        _id: {
-          $oid: "9de52b736c184659b70c4375",
-        },
-        userId: {
-          $oid: "b01342dc577445bc9800bbb6",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "0ac113dabf1d44f98caa329a",
-        },
-        questionId: {
-          $oid: "cab8f4b3472d49c78640b6b2",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 6,
-        successes: 0,
-      },
-    ],
-    myCurrentCourses: [],
-    myCurrentCoursewares: [],
-    myCompletedCourses: [],
-    myCompletedCoursewares: [],
-    createdAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    updatedAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    __v: 0,
-  },
-  {
-    _id: {
-      $oid: "34d5afbb40d64ef9a2a7d603",
-    },
-    name: "User 2",
-    email: "user2@example.com",
-    password: "$2b$10$examplehash",
-    myReviewCards: [
-      {
-        _id: {
-          $oid: "ec5b72519cbb4425957335dd",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "bfb4701e1eb34c51a6ab747b",
-        },
-        questionId: {
-          $oid: "1197909e31dd41b781372c4f",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 3,
-      },
-      {
-        _id: {
-          $oid: "aca703439b434cba8571c435",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "42dc3829c98741348847554e",
-        },
-        questionId: {
-          $oid: "c7e52872445348aaafc4eac7",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 5,
-      },
-      {
-        _id: {
-          $oid: "e12779a77e6345859f66a140",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "b1f1058b5caf4ffdb44cb516",
-        },
-        questionId: {
-          $oid: "8470ab7011a74cd58047eacf",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 2,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "3148b7a4b01746b0a1b244bc",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "e6845359a4b244e885778d69",
-        },
-        questionId: {
-          $oid: "8a90a9a22e0142939e1b15d5",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 10,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "9091c83a1c49438c83e29d0e",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "b1f1058b5caf4ffdb44cb516",
-        },
-        questionId: {
-          $oid: "8470ab7011a74cd58047eacf",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 2,
-        successes: 5,
-      },
-      {
-        _id: {
-          $oid: "cf7e8be0f47f4ef68310f151",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "32f7589f950f477298ccaf3b",
-        },
-        questionId: {
-          $oid: "2478214d7ec64fa380b5963c",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "645f78f97bc14a609120587c",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "0ac113dabf1d44f98caa329a",
-        },
-        questionId: {
-          $oid: "cab8f4b3472d49c78640b6b2",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 10,
-        successes: 3,
-      },
-      {
-        _id: {
-          $oid: "34f7c7b1829049a4a32f3d55",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "b6f9ddb28021471ab0be0b9e",
-        },
-        questionId: {
-          $oid: "1029896111c541ddb85bb661",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "0823ab544fea4245a48bb657",
-        },
-        userId: {
-          $oid: "34d5afbb40d64ef9a2a7d603",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "bfb4701e1eb34c51a6ab747b",
-        },
-        questionId: {
-          $oid: "f0fa0ae314684707a6a7fda1",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 10,
-        successes: 3,
-      },
-    ],
-    myCurrentCourses: [],
-    myCurrentCoursewares: [],
-    myCompletedCourses: [],
-    myCompletedCoursewares: [],
-    createdAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    updatedAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    __v: 0,
-  },
-  {
-    _id: {
-      $oid: "e10d3623578541e1a643b97b",
-    },
-    name: "User 3",
-    email: "user3@example.com",
-    password: "$2b$10$examplehash",
-    myReviewCards: [
-      {
-        _id: {
-          $oid: "1e91249dd16a419f8ad01b5f",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "ac033b12bf454ee39b21bcd4",
-        },
-        questionId: {
-          $oid: "6af323b45c20433086e5b88f",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 8,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "739caa17ab0b426e85e90046",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "eb1321bd92694cbea66b3b4d",
-        },
-        questionId: {
-          $oid: "377ff5a32ee34e5c8ca3680b",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 9,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "4ffbad52124046368cdb6d42",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "641d9a9d1515484dbc594694",
-        },
-        questionId: {
-          $oid: "b247488de11d4b2d88b58d41",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 8,
-        successes: 3,
-      },
-      {
-        _id: {
-          $oid: "c870f4477de2453ba14cca39",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "b4b2a2a432ad4c5f82b76136",
-        },
-        coursewareId: {
-          $oid: "c1ac496500164318b929c985",
-        },
-        questionId: {
-          $oid: "f1001a78be2a49daa407ec9e",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "e656c9642f3a4359bffa6966",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "fda7201018324f5e8dbc9379",
-        },
-        questionId: {
-          $oid: "357384991a60454fa16d0cef",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 1,
-        successes: 1,
-      },
-      {
-        _id: {
-          $oid: "d1d59680ba6e4d00abd21733",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "adc2b4e649c546549f13c006",
-        },
-        questionId: {
-          $oid: "b570fe7f610943309469b6cf",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "8bd5023f44d543efa0ba467a",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "6d9c23e2864640a2a7e76392",
-        },
-        questionId: {
-          $oid: "43bec193a79c4b6b90ec5d17",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 2,
-        successes: 0,
-      },
-      {
-        _id: {
-          $oid: "1f31e9c146e840aea5905178",
-        },
-        userId: {
-          $oid: "e10d3623578541e1a643b97b",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "203394bf8af144b2800d65aa",
-        },
-        questionId: {
-          $oid: "80d0be262f234afaadc270ae",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 1,
-      },
-    ],
-    myCurrentCourses: [],
-    myCurrentCoursewares: [],
-    myCompletedCourses: [],
-    myCompletedCoursewares: [],
-    createdAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    updatedAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    __v: 0,
-  },
-  {
-    _id: {
-      $oid: "a4d624ee69924012a4e40f1a",
-    },
-    name: "User 4",
-    email: "user4@example.com",
-    password: "$2b$10$examplehash",
-    myReviewCards: [
-      {
-        _id: {
-          $oid: "598843329eb54d0e9ee0211f",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "b4b2a2a432ad4c5f82b76136",
-        },
-        coursewareId: {
-          $oid: "adc2b4e649c546549f13c006",
-        },
-        questionId: {
-          $oid: "b570fe7f610943309469b6cf",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 4,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "24a09e2f5d7b45e3a39db38c",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "6be1f5ca22d74c9abf334b66",
-        },
-        questionId: {
-          $oid: "4b83f76e8a744eba9a1c4157",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "2a6092c2b14c46e89583cac4",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "b1f1058b5caf4ffdb44cb516",
-        },
-        questionId: {
-          $oid: "95f25b05eb1e4108848e9af5",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 2,
-        successes: 5,
-      },
-      {
-        _id: {
-          $oid: "164658400b3945b2bc56fcd8",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "c0d4d583bab541b9b55a5f45",
-        },
-        questionId: {
-          $oid: "22015c8ebcd24fa8a263b1c8",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "b1503e7d84ec48619d449357",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "4003f178ff8b401787e09de4",
-        },
-        questionId: {
-          $oid: "fde3d813803f489e98a23e18",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 3,
-      },
-      {
-        _id: {
-          $oid: "e624bf7eded64f87a985e1eb",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "51d6f8068e634d528c7ea322",
-        },
-        questionId: {
-          $oid: "29d502b20b6c4bd3a1920eef",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 7,
-        successes: 3,
-      },
-      {
-        _id: {
-          $oid: "e2e0995f9fea4ac7a67ae974",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "eeac0566a0a04785b25af656",
-        },
-        questionId: {
-          $oid: "2c1d2efd6d074451b31e7d21",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 1,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "10619cb1b52549aba84308bb",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "dfc694ce06de4034b1f71534",
-        },
-        coursewareId: {
-          $oid: "e6845359a4b244e885778d69",
-        },
-        questionId: {
-          $oid: "56f44081a6b049aca39a4ef9",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 6,
-        successes: 5,
-      },
-      {
-        _id: {
-          $oid: "e1ad19d2c8034915b0b62502",
-        },
-        userId: {
-          $oid: "a4d624ee69924012a4e40f1a",
-        },
-        courseId: {
-          $oid: "b4b2a2a432ad4c5f82b76136",
-        },
-        coursewareId: {
-          $oid: "e836fd3cdce24c248a458f66",
-        },
-        questionId: {
-          $oid: "260fd7b9c21545fa991bb0b6",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 3,
-      },
-    ],
-    myCurrentCourses: [],
-    myCurrentCoursewares: [],
-    myCompletedCourses: [],
-    myCompletedCoursewares: [],
-    createdAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    updatedAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    __v: 0,
-  },
-  {
-    _id: {
-      $oid: "20cc73639b354d7087c2b4ee",
-    },
-    name: "User 5",
-    email: "user5@example.com",
-    password: "$2b$10$examplehash",
-    myReviewCards: [
-      {
-        _id: {
-          $oid: "7410a60e5cb74ee584699f9f",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "b4b2a2a432ad4c5f82b76136",
-        },
-        coursewareId: {
-          $oid: "b6f9ddb28021471ab0be0b9e",
-        },
-        questionId: {
-          $oid: "ddddad8ceb3a4fe9a4ba8ef9",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 3,
-        successes: 4,
-      },
-      {
-        _id: {
-          $oid: "a22943666f7f458da07507b7",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "1a9a627de0ee4c63acc4aa3e",
-        },
-        questionId: {
-          $oid: "4b885e93d6a34bb0a26e7024",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 0,
-      },
-      {
-        _id: {
-          $oid: "79abcce078c44043bce8877e",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "709ac14d9cb84698a0720725",
-        },
-        coursewareId: {
-          $oid: "eeac0566a0a04785b25af656",
-        },
-        questionId: {
-          $oid: "7a6b1a07a422464abc2581e7",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 8,
-        successes: 5,
-      },
-      {
-        _id: {
-          $oid: "cb108089465048e3bca594f7",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "7ba27343c11643a2b9e1c53a",
-        },
-        questionId: {
-          $oid: "43d2a121e0ac494e8d2620c6",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 5,
-        successes: 2,
-      },
-      {
-        _id: {
-          $oid: "0d4c1c64e8d74c8abf3a3fc3",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "a1e1084d87dc4502a1dba5f5",
-        },
-        coursewareId: {
-          $oid: "bfb4701e1eb34c51a6ab747b",
-        },
-        questionId: {
-          $oid: "f0fa0ae314684707a6a7fda1",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 4,
-        successes: 1,
-      },
-      {
-        _id: {
-          $oid: "392699699ff44cc3981b111d",
-        },
-        userId: {
-          $oid: "20cc73639b354d7087c2b4ee",
-        },
-        courseId: {
-          $oid: "b9374ca8151541e99d121471",
-        },
-        coursewareId: {
-          $oid: "7ba27343c11643a2b9e1c53a",
-        },
-        questionId: {
-          $oid: "ab2b28e766324c93b420da9a",
-        },
-        nextReviewDate: {
-          $date: "2025-10-01T07:00:00.000Z",
-        },
-        reviews: 9,
-        successes: 0,
-      },
-    ],
-    myCurrentCourses: [],
-    myCurrentCoursewares: [],
-    myCompletedCourses: [],
-    myCompletedCoursewares: [],
-    createdAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    updatedAt: {
-      $date: "2025-09-21T10:48:32.866Z",
-    },
-    __v: 0,
-  },
-];
